+++ tests/conftest.py
@@ -1,0 +1,3 @@
+import os
+os.environ.setdefault("QS_PEPPER", "x" * 32)
+